removed get_history folder
