+++ code/validators.py
@@ -1,4 +1,5 @@
 from config import max_input_length, agent_type
+import uuid
 
 def validate_input(input, request_type):
     """Validates the chat user input. Returns (is_valid, message)."""
@@ -15,3 +16,28 @@ def validate_input(input, request_type):
         request_type = agent_type.GENERIC.value  # <-- fallback string
     
     return True, request_type
+
+def is_valid_uuid(value: str) -> bool:
+    """Check if a string is a valid UUID."""
+    try:
+        uuid.UUID(value)
+        return True
+    except ValueError:
+        return False
+
+
+def validate_session_id(session_id):
+    """Validate session_id and return chat history or create new session."""
+    try:
+        if session_id is None:
+            return False, "session_id is required", 400
+
+        # Validate UUID format
+        if not is_valid_uuid(session_id):
+            return False, "Invalid session_id format", 400
+        
+        return True, "Valid session_id", 200
+
+    except Exception as e:
+        print(f"[Error] {e}")
+        return False, "Internal Server Error", 500
